requirements.txt
fastapi
uvicorn
httpx
openai
python-dotenv
streamlit